backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
openai==1.30.1
edge-tts==6.1.10
moviepy==1.0.3
Pillow==10.3.0
numpy==1.26.4
python-dotenv==1.0.1
aiofiles==23.2.1
